+++ server/runtime.go
@@ -686,7 +686,7 @@ func NewRuntime(ctx context.Context, logger, startupLogger *zap.Logger, db *sql.
 		config.GetSatori().ServerKey,
 		config.GetSatori().SigningKey,
 		config.GetSession().TokenExpirySec,
-		int64(config.GetSatori().HttpTimeoutSec),
+		int64(config.GetSatori().HttpTimeoutMs),
 		!config.GetSatori().CacheDisabled,
 	)