+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project are documented below.
 The format is based on [keep a changelog](http://keepachangelog.com) and this project uses [semantic versioning](http://semver.org).
 
 ## [Unreleased]
+### Changed
+- Adjust unit of measure for Satori client timeout.
+
 ### Fixed
 - Fix migration script to correctly revert console fine-grained access control changes.